+++ moor/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.6 (unreleased)
+- Web support!
+- Date time columns are now comparable
+- Make transactions easier to use: Thanks to some Dart async magic, methods called on your
+  database object in a transaction callback will automatically be called on the transaction object.
+
 ## 1.5.1
 - Fixed an issue where transformed streams would not always update
 - Emit a `INSERT INTO table DEFAULT VALUES` when appropriate. Moor used to generate invalid sql